REFACTOR: Update room row to use decorator instead of temporary column.
This change addresses the issue of having a dependency buried in an unexpected place. Including the building_name column in the rooms table was intended to be a temporary fix to enable sorting on the building name in the short term until we could figure out how to do it properly. Since we already have access to the building associated with the room, we should use that associative data instead of the column that we expect (hopefully) to go away.