FEATURE: Panoramas to room show view
This commit adds the ability to load and display panoramas on the room show view. Panoramas are loaded loaded using Active Storage and rendered using the pannellum javascript library. (** If a panorama is not found in active storage for aparticular room it falls back to using the room_image helper, which in turn falls back to a placeholder image if no image is found.)

The show page renders the new _room_panorama.html.haml partial, which in turn utilized a pannellum stimulus controller.

The pannellum.js library has a css file that is currently loaded in the application.js entrypoint (import 'pannellum/build/pannellum.css'), whcih I don't love (I would prefer we only load the css if we need it), but trying to load it within the pannellum controller was causing issues on my development machine, so there it is for now.

Pannellum requires imagemagick be installed on the machine. For my dev machine (Macintosh with M1 chip) I installed imagemagick using homebrew (brew install imagemagick), which you will need to do as well on your development machines. Hatchbox provided a script to install it on our server instances, which I have already done.

I also created a rooms.sass as a place to put the styles related to rooms. I haven't refactored any of the other sass files to move the rooms specific styling that is scattered about, but now there is a place for them.

Future functionality requests will likely include the ability to copy a 'still' of the image (similar to the pannellum preview) so that Moni can send it in an email, or fax or whatever.