REFACTOR: Skip to content links to allow adding page-specific skip to links.
We added a set_skip_links before action to the ApplicationController which creates a @set_links hash and sets the first element to the id of main and the link text to "Skip to main content".
```
  def set_skip_links
    @skip_links = {main: "Skip to main content"}
  end
```
This default skip_to_content link is applied on every page to allow users to bypass the main navigation links and jump straight to the main content.

We added a skip_links method to the ApplicationHelper. This helper method is included in the application.html.haml template so it runs on each page. It generates a list of links separated the word 'or' (except for the last link) for each item in the @skip_links hash.

```
  # Creates skip to content links from an array of id names
  def skip_links
    sanitize @skip_links.collect { |skip_link| link_to "#{skip_link[1]}", "##{skip_link[0]}", class: "skip-to-content-link" }.join(" <span class='sr-only' > or </span>")
  end
```

To add additional skip_to links on a specific page you need to update the @skip_links hash AND add the corresponding id's to the elements that you want users to be able to jump to.

For example: on the rooms/index page, where we want the users to be able to skip

Update the hash by pushing the additional elements into the @skip_links hash

```
<%- @skip_links = { filters: "Filters", footer: "Footer"} %>
```
And make sure there is a corresponding #Filters id element on the page.

We added a .skip-to-content-link class in the main application.sass file to handle styling of the links. The styling utilizes tailwindcss utilies.

Co-authored-by: Debbie Shih dcjshih@umich.edu
Co-authored-by: Henry Teng heteng@umich.edu